Release v0.0.11
* A new plugin Plugins > Optofluidics > Subtract still defects performs the
median subtraction to the image currently selected, in order to remove the best
we can the static defects we have on images. It can work on arbitrary images.
It operates on the current image, so there is no new image created, but the
pixels of the selected image are modified.  This defect correction is used in
the batch processor on the fly.

* For convenience, there are now two small plugins Plugins > Optofluidics >
Vertical kymograph and Plugins > Optofluidics > Horizontal kymograph. They
generate the kymographs from a single line image, to easily inspect raw data
for defects.

* A new plugin called Plugins > Optofluidics > Configure settings launched a
GUI that allow the edition of the batch processing parameters. I originally
destined it to work with the non-batch application, but here it is.

Signed-off-by: Jean-Yves Tinevez <jeanyves.tinevez@gmail.com>